Fixing build.
musl-crypt-fix.patch has moved. I found it elsewhere, but it was so small that I
figured I'd just include the text so that we don't lose it again.

haproxy 1.5.11-r0 is no longer available from alpine edge but 1.5.12-r0 is, so I
removed the version from the apk-install command. 1.5.12-r0 does not run when
installed, though, because it's missing some symbols. Turns out that openssl-dev
is downgraded during some of the apk-installs that happen in install-nginx and
I needed to explicitly install it from alpine edge in order for everything to
work. This may not be the best fix, I'm open to anything else.